+++ cordapp/src/main/kotlin/net/corda/plugins/CordappPlugin.kt
@@ -31,7 +31,7 @@ class CordappPlugin : Plugin<Project> {
         project.logger.info("Configuring ${project.name} as a cordapp")
 
         if (compareVersions(project.gradle.gradleVersion, MIN_GRADLE_VERSION) < 0) {
-            throw GradleException("Gradle version ${project.gradle.gradleVersion} is below the supported minimum version $MIN_GRADLE_VERSION. Please update Gradle or consider using Gradle wrapper if it is provided with the project. More information about CorDapp build system can be found here: https://docs.corda.net/cordapp-build-systems.html")
+            throw GradleException("Gradle versionId ${project.gradle.gradleVersion} is below the supported minimum versionId $MIN_GRADLE_VERSION. Please update Gradle or consider using Gradle wrapper if it is provided with the project. More information about CorDapp build system can be found here: https://docs.corda.net/cordapp-build-systems.html")
         }
 
         Utils.createCompileConfiguration("cordapp", project)
@@ -50,16 +50,34 @@ class CordappPlugin : Plugin<Project> {
         val task = project.task("configureCordappFatJar")
         val jarTask = project.tasks.getByName("jar") as Jar
         jarTask.doFirst {
-            val (targetPlatformVersion, minimumPlatformVersion) = checkVersionInfo()
             val attributes = jarTask.manifest.attributes
-            attributes["Name"] = cordapp.info.name ?: "${project.group}.${jarTask.baseName}"
-            attributes["Implementation-Version"] = cordapp.info.version ?: project.version
-            attributes["Implementation-Vendor"] = cordapp.info.vendor ?: UNKNOWN
+            var skip = false
+            // Corda 4 attributes support
+            if (cordapp.contract.name != null) {
+                attributes["Cordapp-Contract-Name"] = cordapp.contract.name ?: "${project.group}.${jarTask.baseName}"
+                attributes["Cordapp-Contract-Version"] = parseVersion(cordapp.contract.versionId.toString())
+                attributes["Cordapp-Contract-Vendor"] = cordapp.contract.vendor ?: UNKNOWN
+                attributes["Cordapp-Contract-Licence"] = cordapp.contract.licence ?: UNKNOWN
+                skip = true
+            }
+            if (cordapp.workflow.name != null) {
+                attributes["Cordapp-Workflow-Name"] = cordapp.workflow.name ?: "${project.group}.${jarTask.baseName}"
+                attributes["Cordapp-Workflow-Version"] = parseVersion(cordapp.workflow.versionId.toString())
+                attributes["Cordapp-Workflow-Vendor"] = cordapp.workflow.vendor ?: UNKNOWN
+                attributes["Cordapp-Workflow-Licence"] = cordapp.workflow.licence ?: UNKNOWN
+                skip = true
+            }
+            // Deprecated support (Corda 3)
+            if (skip && cordapp.info.name != null)
+                project.logger.warn("Ignoring deprecated 'info' attributes. Using 'contract' and 'workflow' attributes.")
+            if (!skip && cordapp.info.name != null) {
+                attributes["Name"] = cordapp.info.name ?: "${project.group}.${jarTask.baseName}"
+                attributes["Implementation-Version"] = cordapp.info.version ?: project.version
+                attributes["Implementation-Vendor"] = cordapp.info.vendor ?: UNKNOWN
+            }
+            val (targetPlatformVersion, minimumPlatformVersion) = checkPlatformVersionInfo()
             attributes["Target-Platform-Version"] = targetPlatformVersion
             attributes["Min-Platform-Version"] = minimumPlatformVersion
-            if (attributes["Implementation-Vendor"] == UNKNOWN) {
-                project.logger.warn("CordApp's vendor is \"$UNKNOWN\". Please specify it in \"cordapp.info.vendor\".")
-            }
             if (cordapp.sealing.enabled) {
                 attributes["Sealed"] = "true"
             }
@@ -116,18 +134,32 @@ class CordappPlugin : Plugin<Project> {
         return filteredDeps.toUniqueFiles(runtimeConfiguration) - excludeDeps.toUniqueFiles(runtimeConfiguration)
     }
 
-    private fun checkVersionInfo(): Pair<Int, Int> {
+    private fun checkPlatformVersionInfo(): Pair<Int, Int> {
         // If the minimum platform version is not set, default to 1.
-        val minimumPlatformVersion: Int = cordapp.info.minimumPlatformVersion ?: 1
-        val targetPlatformVersion = cordapp.info.targetPlatformVersion
-                ?: throw InvalidUserDataException("Target version was not set and could not be determined from the project's Corda dependency. Please specify the target version of your CorDapp.")
+        val minimumPlatformVersion: Int = cordapp.minimumPlatformVersion ?: cordapp.info.minimumPlatformVersion ?: 1
+        val targetPlatformVersion = cordapp.targetPlatformVersion ?: cordapp.info.targetPlatformVersion
+                ?: throw InvalidUserDataException("Target versionId was not set and could not be determined from the project's Corda dependency. Please specify the target versionId of your CorDapp.")
         if (targetPlatformVersion < 1) {
-            throw InvalidUserDataException("Target version must not be smaller than 1.")
+            throw InvalidUserDataException("Target versionId must not be smaller than 1.")
         }
         if (targetPlatformVersion < minimumPlatformVersion) {
-            throw InvalidUserDataException("Target version must not be smaller than min platform version.")
+            throw InvalidUserDataException("Target versionId must not be smaller than min platform versionId.")
         }
         return Pair(targetPlatformVersion, minimumPlatformVersion)
+    }
+
+    private fun parseVersion(versionStr: String?): Int {
+        if (versionStr == null)
+            throw InvalidUserDataException("Target versionId not specified. Please specify a whole number starting from 1.")
+        return try {
+            val version = Integer.parseInt(versionStr)
+            if (version < 1) {
+                throw InvalidUserDataException("Target versionId must not be smaller than 1.")
+            }
+            return version
+        } catch (e: NumberFormatException) {
+            throw InvalidUserDataException("Version identifier must be a whole number starting from 1.")
+        }
     }
 
     private fun Iterable<Dependency>.toUniqueFiles(configuration: Configuration): Set<File> {