+++ changelog.md
@@ -4,6 +4,8 @@
 
 ### Version 4.0.37
 
+* `cordapp`: Cordapp naming conventions changed to CordappContract and CordappWorkflow for all attributes (name, version, vendor, licence)
+
 ### Version 4.0.36
 
 * `cordapp`: Fixed the development key for Cordapp signing (incorrect ISO country code in CN).